backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
birdnetlib==0.16.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
firebase-admin==6.6.0
flatbuffers==24.3.25
fonttools==4.55.2
future==1.0.0
gast==0.6.0
geocoder==1.38.1
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h5py==3.12.1
httplib2==0.22.0
idna==3.10
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
namex==0.0.8
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
soundfile==0.12.1
soxr==0.5.0.post1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
watchdog==2.1.9
wheel==0.45.1
wrapt==1.17.0

